pv: If master pv power is not valid, try using sub-values
If one inverter is not communicating, attempt to estimate the PV
power from any other inverter that is working correctly.
This is a hack for a dodgy SMA inverter that sometimes stops
communicating for a couple of days.
The assumption here is that there are only 2 inverters, and they
are roughly equal in output.