db: Add a Fresh method and refactor the value stale checks
Some checks were being done to detect 'stale' data, or data that has
not been updated for some timeout period. This is reasonable, since the source
of the data may have an issue, and it is important to not propagate invalid data.

The method of determining if a data object was fresh was incorrect; the
code that was checking the freshness was running at some interval (e.g each minute), and
was using this interval time to check freshness i.e if the element had not been updated
in the last interval, the object was considered stale. However the data object may be
updated at a different rate (e.g every 2 minutes), meaning the reader may incorrectly consider
the object as stale. Instead the stale timeout should be associated with the object, and
be set according to the update rate of the object (or some sensible default).

New methods have been added to the core data elements for setting the freshness timeout,
and for checking whether this timeout has been exceeded.